refactor the url-converter logic into the urls/ subdirectory
- doesn't change any code, just moves it
- Done in preparation for the introduction layout-based url handlers
